# api/openapi.yaml
openapi: 3.1.0
info:
  title: Prototype API
  version: "1"
servers: []
paths:
  /api/bookings:
    post:
      operationId: createBooking
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
          description: Created
      tags:
      - bookings
  /api/bookings/{id}:
    get:
      operationId: getBooking
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
          description: OK
      tags:
      - bookings
  /api/examples:
    get:
      operationId: listExamples
      parameters:
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseExampleSummaryResponse"
          description: OK
      tags:
      - examples
    post:
      operationId: createExample
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExampleRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExampleResponse"
          description: Created
      tags:
      - examples
  /api/examples/{id}:
    get:
      operationId: getExample
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExampleResponse"
          description: OK
      tags:
      - examples
    put:
      operationId: updateExample
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExampleRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExampleResponse"
          description: OK
      tags:
      - examples
components:
  schemas:
    BookingResponse:
      type: object
      properties:
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
        createdAt:
          type: string
          format: date-time
        customerName:
          type: string
        id:
          type: string
          format: uuid
        licencePlate:
          type: string
        washProgramme:
          type: string
          enum:
          - BASIC
          - COMFORT
          - PREMIUM
      required:
      - appointmentDate
      - appointmentTime
      - createdAt
      - customerName
      - id
      - licencePlate
      - washProgramme
    CreateBookingRequest:
      type: object
      properties:
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
          minLength: 1
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        customerName:
          type: string
          maxLength: 100
          minLength: 1
        licencePlate:
          type: string
          maxLength: 15
          minLength: 1
        washProgramme:
          type: string
          enum:
          - BASIC
          - COMFORT
          - PREMIUM
      required:
      - appointmentDate
      - appointmentTime
      - customerName
      - licencePlate
      - washProgramme
    CreateExampleRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 2000
          minLength: 0
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - CLOSED
        title:
          type: string
          maxLength: 200
          minLength: 0
      required:
      - description
      - status
      - title
    ExampleResponse:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        description:
          type: string
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - CLOSED
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - description
      - id
      - status
      - title
      - updatedAt
    ExampleSummaryResponse:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - CLOSED
        title:
          type: string
      required:
      - createdAt
      - id
      - status
      - title
    PageResponseExampleSummaryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExampleSummaryResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int64
      required:
      - items
      - page
      - size
      - totalItems
    UpdateExampleRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 2000
          minLength: 0
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - CLOSED
        title:
          type: string
          maxLength: 200
          minLength: 0
      required:
      - description
      - status
      - title